+++ src/ota_metadata/legacy2/metadata.py
@@ -57,7 +57,7 @@ from ota_metadata.file_table._orm import (
 from ota_metadata.utils.cert_store import CAChainStore
 from otaclient_common._typing import StrOrPath
 from otaclient_common.common import urljoin_ensure_base
-from otaclient_common.downloader import DownloadInfo
+from otaclient_common.download_info import DownloadInfo
 
 from . import DIGEST_ALG, SUPORTED_COMPRESSION_TYPES
 from .csv_parser import (
